+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+* Fix bug where `GeneratorUtil.overrideMethod(...)` was not passing on the `additionalSuppressions` parameter when generating suppressions.
+
 ### [v0.49](https://github.com/realityforge/proton/tree/v0.49) (2020-05-21) · [Full Changelog](https://github.com/realityforge/proton/compare/v0.48...v0.49)
 
 Changes in this release: